requirements.txt
matplotlib==3.7.0
numpy==1.23.0
Pillow==9.4.0
